mild cleanup
