+++ apps/examples/src/examples/GPTExample/_old-prompt.ts
